Created tests for ShoppingCartOrderService Finished writting tests for UserService Fixed bugs in ShoppingCartOrderService and UserService

+++ bicycle_store_web/Services/ShoppingCartOrderService.cs
@@ -43,6 +43,8 @@ namespace bicycle_store_web.Services
         public bool DeleteShoppingCartOrder(int ShoppingCartOrderId)
         {
             var shoppingCartOrder = _shoppingCartOrderRepo.GetById(ShoppingCartOrderId);
+            if (_shoppingCartOrderRepo.GetById(ShoppingCartOrderId) == null)
+                return true;
             _shoppingCartOrderRepo.Delete(shoppingCartOrder.Id);
             if (_shoppingCartOrderRepo.GetById(ShoppingCartOrderId) != null)
                 return true;
